[refactor] Rename github to _github
