+++ audyn/utils/data/msd/_download.py
@@ -3,7 +3,7 @@ import zipfile
 from typing import Dict, List, Optional, Union
 
 from ... import audyn_cache_dir
-from ...github import download_file_from_github_release
+from ..._github import download_file_from_github_release
 
 
 def download_tags() -> List[Dict[str, str]]:
