+++ tests/package/models/test_ast.py
@@ -16,7 +16,7 @@ from audyn.models.ast import (
 from audyn.models.lextransformer import LEXTransformerEncoderLayer
 from audyn.models.roformer import RoFormerEncoderLayer
 from audyn.modules.vit import PatchEmbedding, PositionalPatchEmbedding
-from audyn.utils.github import download_file_from_github_release
+from audyn.utils._github import download_file_from_github_release
 
 
 def test_official_ast() -> None:
